direct_todo/requirements.txt
django=1.4.3
django-taggit
